+++ src/main/java/com/github/firewolf8385/customitemapi/items/attributes/attributes/AttackSpeedAttribute.java
@@ -17,7 +17,7 @@ public class AttackSpeedAttribute extends ItemAttribute {
     }
 
     @Override
-    public void addedToItem(CustomItem customItem, ItemBuilder item, int value) {
+    public void addedToItem(CustomItem customItem, ItemBuilder item, double value) {
         Attribute attribute = Attribute.GENERIC_ATTACK_SPEED;
 
         switch (customItem.getType()) {
